+++ src/edu/umass/cs/msocket/SocketInfo.java
@@ -58,7 +58,7 @@ public class SocketInfo
   private int                   SocketIdentifier               = -1;
   
   // offset, in seq num, of data read from the current chunk
-  private long                  chunkReadOffsetSeqNum          = 0;
+  private int                  chunkReadOffsetSeqNum          = 0;
   /*
    * how much ahead is the next data message header, before that data is there
    * End seq num of current chunk
@@ -96,7 +96,7 @@ public class SocketInfo
   private long lastNumBytesRecv								   = 0;
   
   // stores the seq num till which data has been sent in handle migration case
-  private long handleMigSeqNum 								   = 0;
+  private int handleMigSeqNum 								   = 0;
   
   // to store byte ranges
   private Vector<ByteRangeInfo> byteInfoVector;
@@ -167,7 +167,7 @@ public class SocketInfo
     }
   }
 
-  public Object byteInfoVectorOperations(int oper, long startSeqNum, int length)
+  public Object byteInfoVectorOperations(int oper, int startSeqNum, int length)
   {
     synchronized (byteInfoVectorMonitor)
     {
@@ -349,7 +349,7 @@ public class SocketInfo
    */
   public int canReadDirect()
   {
-    return (int) (chunkEndSeqNum - chunkReadOffsetSeqNum);
+    return (chunkEndSeqNum - chunkReadOffsetSeqNum);
   }
 
   public synchronized void setchunkEndSeqNum(int s)
@@ -396,7 +396,7 @@ public class SocketInfo
 		lastNumBytesRecv = numBytesRecv;
 	}
   
-  public synchronized void setHandleMigSeqNum(long handleMigSeqNum) {
+  public synchronized void setHandleMigSeqNum(int handleMigSeqNum) {
 		this.handleMigSeqNum = handleMigSeqNum;
 	}
